requirements.txt
docker
numpy
pyyaml